pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "spike-monitor"
version = "1.0.0"
description = "Zero-ops spike detection for Zendesk"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "click>=8.1.7",
]

[project.scripts]
spike-monitor = "spike_monitor.cli:main"
